+++ src/Visits/VisitsClient.php
@@ -13,6 +13,7 @@ use Shlinkio\Shlink\SDK\ShortUrls\Exception\ShortUrlNotFoundException;
 use Shlinkio\Shlink\SDK\ShortUrls\Model\ShortUrlIdentifier;
 use Shlinkio\Shlink\SDK\Tags\Exception\TagNotFoundException;
 use Shlinkio\Shlink\SDK\Visits\Model\OrphanVisit;
+use Shlinkio\Shlink\SDK\Visits\Model\OrphanVisitType;
 use Shlinkio\Shlink\SDK\Visits\Model\Visit;
 use Shlinkio\Shlink\SDK\Visits\Model\VisitsDeletion;
 use Shlinkio\Shlink\SDK\Visits\Model\VisitsFilter;
@@ -154,11 +155,14 @@ readonly class VisitsClient implements VisitsClientInterface
     /**
      * @return VisitsList|OrphanVisit[]
      */
-    public function listOrphanVisitsWithFilter(VisitsFilter $filter): VisitsList
+    public function listOrphanVisitsWithFilter(VisitsFilter $filter, ?OrphanVisitType $type = null): VisitsList
     {
-        return VisitsList::forOrphanVisitsTupleLoader(
-            $this->createVisitsLoaderForUrl('/visits/orphan', $filter->toArray()),
-        );
+        $query = $filter->toArray();
+        if ($type !== null) {
+            $query['type'] = $type->value;
+        }
+
+        return VisitsList::forOrphanVisitsTupleLoader($this->createVisitsLoaderForUrl('/visits/orphan', $query));
     }
 
     /**